Implemented PLI_NPCADD and PLI_NPCDEL. Changed a bunch of variables to unsigned. Instead of sending the player packet buffer once it reaches 4KB, send it if new data would push it over the 64KB limit.  This should prevent frivolous waste caused by the server constantly compressing/sending data that should instead be sent in one huge chunk.

+++ gserver-v2/server/include/TServer.h
@@ -48,7 +48,7 @@ class TServer
 		CLog& getRCLog()						{ return rclog; }
 		
 		TWeapon* getWeapon(const CString& name);
-		TPlayer* getPlayer(const unsigned int id);
+		TPlayer* getPlayer(const unsigned short id);
 		TNPC* getNPC(const unsigned int id);
 
 		TNPC* addNewNPC(const CString& pImage, const CString& pScript, float pX, float pY, TLevel* pLevel, bool pLevelNPC = true);
